fix(sql): 文章流程两个非致命报错 — loadTemplate按name查(非uuid) + MATURE_DAYS cast int
①template-prompt-injector: templateId是逻辑键, 原按.id(uuid)查报'invalid uuid syntax', 改.name
 (注: registry id 与 DB name 命名错配, 仅shunshi-style对得上, 待统一)。
②recipe-learning/asset-performance: c.created_at::date + ${MATURE_DAYS} 参数无类型→'operator not unique: date+unknown', cast ::int。

+++ packages/server/src/services/metrics/asset-performance.ts
@@ -31,7 +31,7 @@ async function aggregateByMetaKey(tenantId: string, contentType: "article" | "vi
         SELECT GREATEST(m.views, m.likes*10 + m.comments*15 + m.shares*20) AS views
         FROM content_metrics m
         WHERE m.content_id = c.id
-          AND m.snapshot_date <= (c.created_at::date + ${MATURE_DAYS})
+          AND m.snapshot_date <= (c.created_at::date + (${MATURE_DAYS})::int)
         ORDER BY m.snapshot_date DESC LIMIT 1
       ) cm ON true
       WHERE c.tenant_id = ${tenantId}
