+++ packages/server/src/services/metrics/recipe-learning.ts
@@ -42,7 +42,7 @@ export async function getRecipeWeights(tenantId: string): Promise<RecipeWeights 
         SELECT GREATEST(m.views, m.likes*10 + m.comments*15 + m.shares*20) AS views
         FROM content_metrics m
         WHERE m.content_id = c.id
-          AND m.snapshot_date <= (c.created_at::date + ${MATURE_DAYS})
+          AND m.snapshot_date <= (c.created_at::date + (${MATURE_DAYS})::int)
         ORDER BY m.snapshot_date DESC LIMIT 1
       ) cm ON true
       WHERE c.tenant_id = ${tenantId}
